Reuse the logic in `mypy_test.py` to get rid of some hacks
